+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Added `take` method to `Repository` to allow fetching resources without id.
 * Added meta to `ItemInterface`.
 N.B. This is a breaking change if you implement the `ItemInterface` yourself instead of using the supplied `Item`.
+* Added `Jsonapi` class.
 
 ### Changed
 
@@ -18,10 +19,15 @@ N.B. This is a breaking change if you implement the `ItemInterface` yourself ins
 N.B. This is a breaking change if you use meta on errors.
 * `DocumentInterface::getMeta()` now returns a `Meta` instance instead of a plain array. If no meta is present, it returns `null`. All implementations have been updated to reflect these changes.
 N.B. This is a minor breaking change if you use meta on documents.
+* `DocumentInterface::getJsonapi()` now returns a `Jsonapi` instance instead of a plain array. If no jsonapi is present, it returns `null`. All implementations have been updated to reflect these changes.
 
 ### Removed
 
 * Removed `ErrorMeta` class in favor of generic `Meta` class.
+
+### Fixed
+
+* Fixed parsing of [JSON:API object](https://jsonapi.org/format/#document-jsonapi-object) in document.
 
 ## [0.14.0] - 2019-01-23